requirements.txt
requests
pandas
streamlit
matplotlib
apache-airflow